+++ modules/interpreter/src/cpp/operators/PowerOperator.cpp
@@ -9,8 +9,10 @@
 //=============================================================================
 #include "Evaluator.hpp"
 #include "Operators.hpp"
-#include "OverloadBinaryOperator.hpp"
 #include "DotPower.hpp"
+#include "OverloadHelpers.hpp"
+#include "FindCommonType.hpp"
+#include "IEEEFP.hpp"
 //=============================================================================
 namespace Nelson {
 //=============================================================================
@@ -18,19 +20,89 @@ ArrayOf
 Evaluator::powerOperator(AbstractSyntaxTreePtr t)
 {
     callstack.pushID((size_t)t->getContext());
-    ArrayOf retval = this->powerOperator(expression(t->down), expression(t->down->right));
+    ArrayOfVector args;
+    args << expression(t->down);
+    args << expression(t->down->right);
+    ArrayOf retval = this->powerOperator(args);
     callstack.popID();
     return retval;
 }
 //=============================================================================
 ArrayOf
-Evaluator::powerOperator(const ArrayOf& A, const ArrayOf& B)
+Evaluator::powerOperator(const ArrayOfVector& args)
 {
-    ArrayOf retval;
-    bool needToOverload;
-    ArrayOf res = DotPower(A, B, needToOverload);
+    const std::string functionName = POWER_OPERATOR_STR;
+    std::string commonTypeName = NLS_UNKNOWN_STR;
+    NelsonType commonType = NLS_UNKNOWN;
+    bool isSparse = false;
+    bool isComplex = false;
+
+    ArrayOf res;
+    if (FindCommonType(args, commonType, isSparse, isComplex, commonTypeName)) {
+        bool overloadWasFound = false;
+        res = callOverloadedFunction(this,
+            NelsonConfiguration::getInstance()->getOverloadLevelCompatibility(), args, functionName,
+            commonTypeName, commonType, overloadWasFound);
+        if (overloadWasFound) {
+            return res;
+        }
+    }
+
+    bool needToOverload = false;
+    ArrayOf A(args[0]);
+    ArrayOf B(args[1]);
+    if (A.getDataClass() != B.getDataClass()) {
+        if (A.isIntegerType()) {
+            bool isCompatible = (B.getDataClass() == NLS_DOUBLE) && B.isScalar();
+            if (!isCompatible) {
+                Error(_W("Integers can only be combined with integers of the same class, or scalar "
+                         "doubles."));
+            }
+            A.promoteType(commonType);
+
+            auto* ptrB = (double*)B.getDataPointer();
+            indexType elementCount = B.getElementCount();
+            bool allIntegerValue = IsIntegerForm(ptrB, elementCount);
+            if (!allIntegerValue) {
+                Error(_W("Positive integral powers expected."));
+            }
+
+            B.promoteType(commonType);
+
+        } else if (B.isIntegerType()) {
+            bool isCompatible = (A.getDataClass() == NLS_DOUBLE) && A.isScalar();
+            if (!isCompatible) {
+                Error(_W("Integers can only be combined with integers of the same class, or scalar "
+                         "doubles."));
+            }
+            A.promoteType(commonType);
+            B.promoteType(commonType);
+        } else {
+            if (commonType <= NLS_CHAR) {
+                NelsonType _commonType = commonType;
+                if (_commonType == NLS_DOUBLE && isComplex) {
+                    _commonType = NLS_DCOMPLEX;
+                }
+                if (_commonType == NLS_SINGLE && isComplex) {
+                    _commonType = NLS_SCOMPLEX;
+                }
+                if (_commonType == NLS_CHAR) {
+                    _commonType = NLS_DOUBLE;
+                }
+                A.promoteType(_commonType);
+                B.promoteType(_commonType);
+            }
+        }
+    }
+    res = DotPower(A, B, needToOverload);
+
     if (needToOverload) {
-        OverloadRequired(POWER_OPERATOR_STR);
+        bool overloadWasFound = false;
+        res = callOverloadedFunction(this, NLS_OVERLOAD_ALL_TYPES, args, functionName,
+            commonTypeName, commonType, overloadWasFound);
+        if (!overloadWasFound) {
+            OverloadRequired(functionName);
+        }
     }
     return res;
 }